update: refactored client actions to manageable smaller modules, updated loader in admin profile page, added forgot password ui
